+++ ontolearner/learner/taxonomy_discovery/__init__.py
@@ -15,4 +15,5 @@
 from .alexbek import AlexbekCrossAttnLearner
 from .rwthdbis import RWTHDBISSFTLearner
 from .sbunlp import SBUNLPFewShotLearner
+from .semanticswingers import SemanticSwingersTaxonomyLearner
 from .skhnlp import SKHNLPSequentialFTLearner, SKHNLPZSLearner